update
kode kelas added,
code cleanup,
minor improvements.

+++ app/Database/Migrations/2023-07-19-094221_NilaiSiswa.php
@@ -26,7 +26,7 @@ class NilaiSiswa extends Migration
                 'unsigned' => true,
             ],
             'nilai' => [
-                'type' => 'INT',
+                'type' => 'FLOAT',
                 'constraint' => 11,
             ],
             'id_kelas' => [
